Workshop/requirements.txt
pydantic
datasets
pyarrow
wandb
peft>=0.5.0
trl
bitsandbytes>=0.42.0
huggingface-hub
torch>=2.0.0
transformers>=4.30.0
gradio>=3.50.0
accelerate>=0.20.0
ipykernel
ipywidgets
google-genai
python-dotenv
matplotlib
litellm
pandas
plotly
tqdm
html2text
pytest
pytest-mock
pytest-asyncio
beautifulsoup4